64bit to 32bit

+++ src/framework.rs
@@ -2,7 +2,7 @@ use crate::def::Command;
 use crate::neighborhood::Neighborhood;
 
 pub trait IState {
-    fn get_score(&self, progress: f64) -> f64;
+    fn get_score(&self, progress: f32) -> f32;
     fn perform_command(&mut self, command: &Command) -> Vec<Command>;
     fn reverse_command(&mut self, command: &Command);
 }
@@ -13,10 +13,10 @@ pub trait INeighborhoodSelector {
 }
 
 pub trait IOptimizer {
-    fn update_temp(&mut self, progress: f64);
-    fn should_adopt_new_state(&self, score_diff: f64) -> bool;
+    fn update_temp(&mut self, progress: f32);
+    fn should_adopt_new_state(&self, score_diff: f32) -> bool;
 }
 
 pub trait ISolver {
-    fn solve(&mut self, time_limit: f64);
+    fn solve(&mut self, time_limit: f32);
 }
